fix: cannot unmarshal number 7632548910 into Go struct field .chats.context.value.id of type int32

+++ messenger/messenger.gen.go
@@ -270,7 +270,7 @@ type Chat struct {
 		Type  *string `json:"type,omitempty"`
 		Value *struct {
 			// Id ID объявления
-			Id *int32 `json:"id,omitempty"`
+			Id *int64 `json:"id,omitempty"`
 
 			// Images Изображения с карточки объявления
 			Images *struct {
